Image creation: Make use of the mkfs.jffs2 "sortfiles" option.
Due to JFFS2 garbage collection and wear leveling, the kernel and cfe
files may be relocated. This causes problems for cfe.

This is a TEMPORARY fix that uses mkfs.jffs2 to allocate the kernel
and cfe to their own erase blocks thus making them less likely for
garbage collection (however eventually they wil be moved so this
only postpones the problem).