Kill JFFS2 garbage collector before starting upgrade, refs #5702
The JFFS2 garbage collector and the upgrade process competes for the brcmnand
status (~lock). If the GC holds the lock and executes on the CPU that is shut
down, the upgrade process will fail or wait forever for the lock.

This patch kills jffs2_gcd_mtd* before launching the brcm_fw_tool.