permngr: optimize use of I2C.
We don't want to do I2C commands unless 100% needed.
I2C is emulated over gpio and is slow and takes CPU.

here we use a cached value of state and only do the actual write if
new state is different.

change has effect on reset and that now needs to restore state to the leds
and not use the cached state.