permngr: Add new top level leds.
We need this to control the leds for some eth ports where the
phy is not in control.

Try to avoid abusing the led action states to control what led is lit.

It makes it very hard to read code when it say example: set state notice
followed by state error and what we really do is turn on two different
leds.