log everything at boot.
This capture all output during boot even things going to stdout stderr
into the syslog.

Almost that is. /sbin/init is run with stdout/stderr to the console and
syslog is not started until S10_boot is run and quite a bit into that file
so there is some output that could be missed.  I did not want init to run
redirected to /dev/kmsg.

To improve the printouts boot scripts should pipe to /dev/kmsg with
&> /dev/ksmsg or 2>&1 >/dev/kmsg. That way we don't use logger and removes
the extra empty line that it produces. Alternatively use logger and not
also print to stderr as writing directly to /dev/kmsg means we lose the
information on what script is printing.

Also some things don't demonize properly. after boot
ps | grep logger

should not show any logger process now it is

  822 root      1568 S    logger -s -p 6 -t sysinit
 3058 root      1568 S    logger -s -p 6 -t sysinit_/etc/rc.d/S20swmdk
 8957 root      1568 S    logger -s -p 6 -t sysinit_/etc/rc.d/S97dect
 9319 root      1568 S    logger -s -p 6 -t sysinit_/etc/rc.d/S99ice-client

refs: #6823