base_files/upgrade: Dynamically determine /dev/ubi_ctrl minor
UBI uses the "misc" device major (=10) for the /dev/ubi_ctrl device
node. As this major is shared with other drivers, the minor number
assigned to UBI may change depending on kernel config, loaded
drivers etc. Normally this is taken care of by the device manager
but when creating the minimal rootfs for partition upgrade it
becomes an issue.