reboot: detect first boot. detect default reset.
if REBOOT_REASON env variable is set when reboot is called
that string is used in the next boot to indicate why the
board rebooted. for now only used for defaultreset.

First boot is detected if the env variable FIRST_BOOT is set.

if no reason is found we default to think it is watchdog.