preinit: avoid sourcing /lib/iopsys.sh
boot procedure is as follows.

kernel starts /etc/preinit.

[ -z "$PREINIT" ] && exec /sbin/init

in preinit will at boot result in /sbin/init to replace /etc/preinit
as the process on PID 1.

/sbin/init is procd and it will start /etc/preinit again as a sub process
on a different PID this time with PREINIT set to 1 so that we can get past

[ -z "$PREINIT" ] && exec /sbin/init

and run the rest of the scripts.

this is just avoiding to load
/lib/iopsys.sh and run iopsys_upgrade_handling twice.
may save a few hundred milliseconds.